Moved files to library package
