+++ src/library/Variables.java
@@ -1,4 +1,4 @@
-package main;
+package library;
 
 public class Variables {
 	// Dynamic global variables
